Consolidate saved prompt commands
