+++ extensions/multi-ai-chat/CHANGELOG.md
@@ -4,5 +4,5 @@
 
 - Open one prompt in ChatGPT, Claude, Grok, and Perplexity using existing browser sessions.
 - Choose zero to five tabs for each supported provider.
-- Create reusable prompt presets with named arguments and Raycast Quicklinks.
+- Create reusable saved prompts with named arguments and Raycast Quicklinks.
 - Select a preferred browser or use the macOS system default.
